packets_misc.c
