+++ Makefile
@@ -256,6 +256,7 @@ obj/music_player.o \
 obj/net_game.o \
 obj/net_sync.o \
 obj/packets.o \
+obj/packets_misc.o \
 obj/player_compchecks.o \
 obj/player_compevents.o \
 obj/player_complookup.o \
