+++ src/packets.h
@@ -185,6 +185,21 @@ enum TbPacketAddValues {
     PCAdV_Unknown80         = 0x80, //!< Seem unused
 };
 
+enum ChecksumKind {
+    CKS_Action = 0,
+    CKS_Players,
+    CKS_Creatures_1,
+    CKS_Creatures_2,
+    CKS_Creatures_3,
+    CKS_Creatures_4,
+    CKS_Creatures_5,  // Heroes
+    CKS_Creatures_6,  // Neutral
+    CKS_Things, //Objects, Traps, Shots etc
+    CKS_Effects,
+    CKS_Rooms,
+    CKS_MAX
+};
+
 #define PCtr_LBtnAnyAction (PCtr_LBtnClick | PCtr_LBtnHeld | PCtr_LBtnRelease)
 #define PCtr_RBtnAnyAction (PCtr_RBtnClick | PCtr_RBtnHeld | PCtr_RBtnRelease)
 #define PCtr_HeldAnyButton (PCtr_LBtnHeld | PCtr_RBtnHeld)
@@ -223,6 +238,12 @@ struct PacketSaveHead { // sizeof=0xF (15)
     TbBool chksum_available; // if needed, this can be replaced with flags
 };
 
+struct PacketEx
+{
+    struct Packet packet;
+    TbBigChecksum sums[CKS_MAX];
+};
+
 #pragma pack()
 /******************************************************************************/
 /******************************************************************************/
@@ -235,8 +256,8 @@ void set_players_packet_control(struct PlayerInfo *player, unsigned long flag);
 unsigned char get_players_packet_action(struct PlayerInfo *player);
 void unset_packet_control(struct Packet *pckt, unsigned long flag);
 void unset_players_packet_control(struct PlayerInfo *player, unsigned long flag);
-void set_players_packet_position(struct PlayerInfo *player, long x, long y);
-short set_packet_pause_toggle(void);
+void set_players_packet_position(struct Packet *pckt, long x, long y, unsigned char context);
+void set_packet_pause_toggle(void);
 TbBool process_dungeon_control_packet_clicks(long idx);
 TbBool process_players_dungeon_control_packet_action(long idx);
 void process_players_creature_control_packet_control(long idx);
